Reality2 (Reality2.AI v0.1.0)

Reality2 Sentient Agent (Sentant) based Platform with plugin architecture for Intuitive Spatial Computing supporting Assistive Technologies.

This is the main module for managing Sentants, Swarms and Plugins on a Node and in a Cluster. Primarily, this will not be accessed directly, but rather through the GraphQL API Reality2.Web.

Plugins

ai.reality2.vars - A plugin for managing variables on a Sentant.

ai.reality2.geospatial - Coming Soon - A plugin for managing geospatial location and context on a Sentant.

ai.reality2.pathing - Coming Soon - A plugin for managing pathing on a node, cluster and globally - interfaces with the Pathing Name System (PNS) and ensures Sentant Global Uniqueness and Addressability.

Author

Dr. Roy C. Davies

roycdavies.github.io

Functions

test_explugin()