Reality2.Sentants (Reality2.AI v0.1.0)

Module for creating and managing Sentants, and the DynamicSupervisor that manages them.

When a Sentant is created, it is given a unique ID, and a name. The name is unique on the node, but not in the world. Upon creation, the Sentant is sent the init event, which is handled by the Sentant's Automations.

If a Sentant with the same name or ID already exists on the Node, then it is not created again, but redefined and restarted.

Sentants are immutable, so they cannot be changed once created. To change a Sentant, it must be reloaded. However, some data in plugins may change (such as Reality2.AiReality2Vars), and this is handled by the plugin itself. Further, the state(s) of the Sentant Automations can change.

Author

Dr. Roy C. Davies

roycdavies.github.io

Types

sentant_definition()
(opaque)

@opaque sentant_definition()

The definition of a Sentant is a string containing the definition of the Sentant in YAML format. See the definition of YAML.Sentant.

sentant_name_or_uuid()

@type sentant_name_or_uuid() :: %{id: Reality2.Types.uuid()} | %{name: String.t()}

Each Sentant can be referred to by either its name or its ID. The name of a Sentant is unique on the node, but not in the world. This is used in pathing.

Functions

create(definition_map)

@spec create(
  (definition :: sentant_definition())
  | (definition_map :: Reality2.Types.sentant())
) ::
  {:ok, String.t()} | {:error, :definition}

Create a new Sentant and return the result of the operation with the pid of the new Sentant, or an appropriate error.

Parameters

definition - A string containing the definition of the Sentant to be created in YAML format.

(or) 'definition_map' - A map as created from a YAML definition.

Returns

{:ok, id} - The Sentant was created.

{:error, :definition} if the definition is invalid.

delete(name_or_uuid)

@spec delete(name_or_uuid :: sentant_name_or_uuid()) ::
  {:ok, Reality2.Types.uuid()}
  | {:error, :name}
  | {:error, :existance}
  | {:error, :id}

Delete a Sentant and return the result of the operation with the pid of the deleted Sentant, or an appropriate error.

Parameters

name_or_uuid - The name or ID of the Sentant to be deleted as a map containing either a :name or :id key.

Returns

{:ok, id} - The Sentant was deleted.

{:error, :name} if the Sentant with that name does not exist on this node.

{:error, :id} if the Sentant with that ID does not exist on this node.

{:error, :existance} if the Sentant with that ID or name does not exist on this node.

Example

Reality2.Sentants.delete(%{:name => "my_sentant"})

read(name_or_uuid, command)

@spec read(name_or_uuid :: sentant_name_or_uuid(), command :: :state | :definition) ::
  {:ok, map()} | {:error, :name} | {:error, :existance} | {:error, :id}

Read something from an existing Sentant - determined by the command. The result will depend on the command.

Parameters

name_or_uuid - The name or ID of the Sentant to be read from as a map containing either a :name or :id key.

command - The command to be executed on the Sentant, which must be either :state or :definition.

Returns

{:ok, definition} - The Sentant was read.

{:error, :existance} if the Sentant with that ID or name does not exist on this node.

{:error, :invalid} if the parameter is invalid.

Example

case Reality2.Sentants.read(%{:name => "my_sentant"}, :definition) do
{:ok, definition} ->
  # Do something with the definition
{:error, :existance} ->
  # Sentant does not exist
{:error, :invalid} ->
  # Invalid parameter
end

Reality2.Sentants.read(%{id: "123e4567-e89b-12d3-a456-426614174000"}, :state)

sendto(name_or_uuid, message_map)

@spec sendto(name_or_uuid :: sentant_name_or_uuid(), message :: map()) ::
  {:ok} | {:error, :name} | {:error, :existance} | {:error, :id}

Send a message to the named Sentant if it exists and return the result of the operation, or an appropriate error. This is an asynchronous operation.

Parameters

name_or_uuid - The name or ID of the Sentant to have the message sent to as a map containing either a :name or :id key.

message - The message to be sent, which must contain a :command string and optionally a :parameters map, and a :passthrough map.

Returns

{:ok} - The message was sent.

{:error, :name} if the Sentant with that name does not exist on this node.

{:error, :id} if the Sentant with that ID does not exist on this node.

{:error, :existance} if the Sentant with that ID or name does not exist on this node.

Example

Reality2.Sentants.sendto(%{:name => "my_sentant"}, %{event: "turn_on", delay: 1000})

sendto_all(message_map)

@spec sendto_all(message :: map()) :: {:ok, integer()}

Send a message to all Sentants. This is an asynchronous operation, so the result is always {:ok}.

Parameters

message - The message to be sent, which must contain a :command string and optionally a :parameters map, and a :passthrough map.

Returns

{:ok, num_sentants} - The number of Sentants that the message was sent to.

Example

Reality2.Sentants.sendto_all(%{event: "turn_on"})