Reality2.Types (Reality2.AI v0.1.0)

This is a Sentant definition template. Follow this guide to define a Sentant Template or a specific Sentant.

NOTE: This is a work in progress and is subject to change.

Version: 0.0.1

Date: 2023.11.18

Author

Dr. Roy C. Davies

roycdavies.github.io

Types

action()

@type action() :: %{plugin: plugin(), command: String.t(), parameters: map()}

The action definition of an automation.

YAML

action:
  # Optional plugin to use for this action
  plugin: !plugin

  # The command to execute
  command: !str

  # The parameters of the command in JSON format
  parameters: { !str: !json }

author()

@type author() :: %{id: String.t(), name: String.t(), email: String.t()}

A description of who made or owns the Sentant.

YAML

author:
  # The author's ID
  id: !guid

  # The author's name
  name: !str

  # The author's email
  email: !str

automation()

@type automation() :: %{
  name: String.t(),
  description: String.t(),
  transitions: [transition()]
}

A sentant automation definition.

Each Sentant has a default Automation that determines it's current status, such as active and shadow, and which is preprogrammed to handle the init, join_cluster, leave_cluster, connect, disconnect, go_shadow, go_active, shadows, active and transfer events, and change the Sentant state accordingly. To access the current state of that default Automation, use the automation name "_" (underscore).

YAML

automation:
  # The name of the automation (must be unique within the Sentant)
  name: !str

  # The description of the automation
  description: !str

  # The transitions of this automation
  transitions: [ !transition ]

input_event()

@type input_event() :: %{event: String.t(), parameters: map(), passthrough: map()}

An event definition sent to a Sentant.

YAML

input_event:
  # The name of the event (to match the transition event)
  event: !str

  # The parameters of the event in JSON format
  parameters: { !str: !json }

plugin()

@type plugin() :: %{
  name: String.t(),
  url: String.t(),
  method: String.t(),
  headers: map(),
  body: map(),
  output: map(),
  version: String.t()
}

Plugins are used to add functionality to Sentants beyond the built-in actions.

The name is used to specify exactly how the plugin works and is defined separately in a plugin definition file.

YAML

plugin:
  # The name of the plugin in reverse domain notation, eg: ai.reality2.storage
  name: !str

  # The version of the plugin, eg: "0.1.0"
  version: !str

sentant()

@type sentant() :: %{
  id: uuid(),
  name: String.t(),
  version: String.t(),
  class: String.t(),
  data: map(),
  binary: map(),
  tags: [String.t()],
  keywords: [String.t()],
  description: String.t(),
  author: author(),
  plugins: [plugin()],
  automations: [automation()],
  states: [stored_state()],
  status: status()
}

The definition of a Sentant.

YAML

sentant:
  # The guid of the Sentant - for a Sentant Template, this is left empty
  # Loading a Sentant with an empty guid will cause a new Sentant to be created
  # If a guid is provided, the Sentant will be loaded and checked for uniqueness before being allowed
  # to function
  id: !guid

  # The name of the Sentant - best to keep short and simple as it will become part of the pathing.
  # Avoid using spaces or special characters.
  # The dot '.' character has a special meaning in the pathing, so cannot be used in the name.
  # The name is case insensitive, so "MySentant" and "mysentant" are the same.
  # New Sentants with the same name on a Node will have a unique 6 digit hexcode hash appended to the
  # end of the name to ensure uniqueness on the Node
  name: !str

  # The version of the Sentant, eg: "0.1.0".  When used for a Sentant Template, this is the minimum
  # version of Sentant that can be used with this template.
  # When used for a specific Sentant, this is the version of the Sentant.
  version: !str

  # The class of the Sentant - in reverse domain notation, eg: ai.reality2.sentant.default
  class: !str

  # The initial (immutable) data of the Sentant in key/value pairs of JSON formatted data
  data: { !str: !json }

  # The initial (immutable) binary data of the Sentant in key/value pairs of binary data, eg images,
  # audio, video, etc
  binary: { !str: !binary }

  # A list of strings that may be used to group Sentants together in a Node
  keywords: [ !str ]

  # A list of strings that may be used to search for Sentants in a Node
  tags: [ !str ]

  # The description of the Sentant
  description: !str

  # The author of the Sentant
  author: !author

  # Sentant automations
  automations: [ !automation ]

  # Current stored states of the Sentant (empty if Sentant Template)
  states: [ !stored_state ]

  # The status of the Sentant
  status: !str

status()

@type status() :: String.t()

The current status of this Sentant on this node.

YAML

"active" | "shadow" | "inactive" | "unchecked" | "unknown"

stored_state()

@type stored_state() :: %{name: String.t(), state: String.t()}

A stored state of an Automation.

YAML

stored_state:
  # The name of the automation
  name: !str

  # The state of the automation
  state: !str

swarm()

@type swarm() :: %{
  name: String.t(),
  class: String.t(),
  description: String.t(),
  author: author(),
  version: String.t(),
  sentants: [sentant()]
}

A group of Sentant Templates that work together to achieve a common goal.

YAML

swarm:
  # The name of the swarm
  name: !str

  # The class of the swarm - in reverse domain notation, eg: ai.reality2.swarm.default
  class: !str

  # The description of the swarm
  description: !str

  # The author of the swarm
  author: !author

  # The version of the swarm, eg: "0.1.0"
  version: !str

  # The Sentant Templates that make up the swarm
  sentants: [ !sentant ]

transition()

@type transition() :: %{
  from: String.t(),
  event: String.t(),
  to: String.t(),
  actions: [action()]
}

An automation transition action definition.

When an event is sent to a Sentant, the parameters are sent to the first action, and the result of that action is sent to the next action, and so on.

Paremeters passed in may be passed on to the next action, or may be modified, or new parameters may be added. Note that all automations start in the "start" state, and the first event sent is the "init" event. There are some events that are reserved for the Node and Sentant system, and these are:

State Changing Events

init - sent when the Sentant is first created (as above)

join_cluster - sent when the Node joins a cluster

leave_cluster - sent when the Node leaves a cluster

connect - sent when the Node the Sentant is on connects to another Node or Cluster (this is different from joining a cluster)

disconnect - sent when the Node the Sentant is on disconnects from another Node or Cluster (this is different from leaving a cluster)

go_shadow - sent when the Sentant is moved to shadow status

go_active - sent when the Sentant is moved to active status

Query Events

get_shadows - returns a list of the paths or IP address and path to all (known) shadow Sentants

get_active - returns the path or IP address and path to the currently active Sentant

get_states - returns the current states of the Sentant

Other Events

do_transfer - transfer control of the Sentant to one of the shadow Sentants (thus becoming a shadow Sentant itself and sending a go_shadow event)

YAML

transition:
  # State to match (* = any)
  from: !str

  # Event to match
  event: !str

  # New state to transition to (* = no change)
  to: !str

  # The actions to perform when transition occurs
  actions: [ !action ]

uuid()
(opaque)

@opaque uuid()

A universal ID.

Functions

action()

author()

automation()

input_event()

plugin()

plugin_defaults()

sentant()

status()

stored_state()

swarm()

transition()

validate(data, typedef)